Revert "회원 정보를 관리할 수 있는 페이지를 생성한다."
