--- src/pages/mypage/Mypage.tsx
@@ -1,14 +1,0 @@
-import { Outlet } from 'react-router-dom';
-import SideBar from '../../components/SideBar';
-import { MainContainer } from '../../styles/commonComponents';
-
-const Mypage = () => {
-  return (
-    <MainContainer>
-      <SideBar />
-      <Outlet />
-    </MainContainer>
-  );
-};
-
-export default Mypage;
